+++ app/src/main/java/org/hyperledger/besu/services/BlockSimulatorServiceImpl.java
@@ -172,6 +172,7 @@ public class BlockSimulatorServiceImpl implements BlockSimulationService {
           new BlockSimulationParameter.BlockSimulationParameterBuilder()
               .blockStateCalls(List.of(blockStateCall))
               .validation(true)
+              .enforceConsensusGasLimitCaps(true)
               .fakeSignature(fakeSignature)
               .build();