+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/transaction/BlockSimulationParameter.java
@@ -41,6 +41,7 @@ public class BlockSimulationParameter {
 
   final List<? extends BlockStateCall> blockStateCalls;
   private final boolean validation;
+  private final boolean enforceConsensusGasLimitCaps;
   private final boolean traceTransfers;
   private final boolean returnFullTransactions;
   private final boolean returnTrieLog;
@@ -63,6 +64,7 @@ public class BlockSimulationParameter {
     this(
         blockStateCalls,
         validation,
+        false,
         traceTransfers,
         returnFullTransactions,
         returnTrieLog,
@@ -72,6 +74,7 @@ public class BlockSimulationParameter {
   public BlockSimulationParameter(
       final List<? extends BlockStateCall> blockStateCalls,
       final boolean validation,
+      final boolean enforceConsensusGasLimitCaps,
       final boolean traceTransfers,
       final boolean returnFullTransactions,
       final boolean returnTrieLog,
@@ -79,6 +82,7 @@ public class BlockSimulationParameter {
     checkNotNull(blockStateCalls);
     this.blockStateCalls = blockStateCalls;
     this.validation = validation;
+    this.enforceConsensusGasLimitCaps = enforceConsensusGasLimitCaps;
     this.traceTransfers = traceTransfers;
     this.returnFullTransactions = returnFullTransactions;
     this.returnTrieLog = returnTrieLog;
@@ -91,6 +95,10 @@ public class BlockSimulationParameter {
 
   public boolean isValidation() {
     return validation;
+  }
+
+  public boolean isEnforceConsensusGasLimitCaps() {
+    return enforceConsensusGasLimitCaps;
   }
 
   public boolean isTraceTransfers() {
@@ -204,6 +212,7 @@ public class BlockSimulationParameter {
   public static class BlockSimulationParameterBuilder {
     private List<? extends BlockStateCall> blockStateCalls = List.of();
     private boolean validation = false;
+    private boolean enforceConsensusGasLimitCaps = false;
     private boolean traceTransfers = false;
     private boolean returnFullTransactions = false;
     private boolean returnTrieLog = false;
@@ -218,6 +227,12 @@ public class BlockSimulationParameter {
 
     public BlockSimulationParameterBuilder validation(final boolean validation) {
       this.validation = validation;
+      return this;
+    }
+
+    public BlockSimulationParameterBuilder enforceConsensusGasLimitCaps(
+        final boolean enforceConsensusGasLimitCaps) {
+      this.enforceConsensusGasLimitCaps = enforceConsensusGasLimitCaps;
       return this;
     }
 
@@ -246,6 +261,7 @@ public class BlockSimulationParameter {
       return new BlockSimulationParameter(
           blockStateCalls,
           validation,
+          enforceConsensusGasLimitCaps,
           traceTransfers,
           returnFullTransactions,
           returnTrieLog,